pyproject.toml
[project]
name = "themomart-controller"
version = "0.1.0"
description = "Raspberry Pi interface to Thermomart PID-RS kiln controller"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "minimalmodbus>=2.1",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "pydantic>=2.0",
    "luma.oled>=3.13",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "mypy>=1.14",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "httpx>=0.28",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.14"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "backend/modbus/real_controller.py",
    "backend/services/display.py",
]
